Summary for projection pursuit based on Gaussian mixtures and evolutionary algorithms for data visualisation

Summary method for objects of class 'ppgmmga'.

Usage

# S3 method for ppgmmga
summary(object, check = (object$approx != "none"), ...)
                
# S3 method for summary.ppgmmga
print(x, digits = getOption("digits"), ...)

Arguments

object

An object of class 'ppgmmga' as returned by ppgmmga.

check

A logical value specifying whether or not a Monte Carlo negentropy approximation check should be performed. By default is FALSE for exact negentropy calculation and TRUE for approximated negentropy.

x

An object of class summary.ppgmmga.

digits

The number of significant digits.

...

Further arguments passed to or from other methods.

Value

The summary function returns an object of class summary.ppgmmga which can be printed by the corresponding print method. A list with the information from the ppgmmga algorithm is returned.

If the optional argument check = TRUE then the value of negentropy is compared to the Monte Carlo negentropy calculated for the same optimal projection basis selected by the algorithm. By default, it allows to check if the value returned by the employed approximation is closed to the Monte Carlo approximation of to the "true" negentropy. The ratio between the approximated value returned by the algorithm and the value computed with Monte Carlo is called Relative Accuracy. Such value should be close to 1 for a good approximation.

Author

Serafini A. srf.alessio@gmail.com
Scrucca L. luca.scrucca@unipg.it

See also